fix: move components that where capitalized
