+++ packages/@dcl/inspector/src/components/Controls/index.ts
@@ -1,0 +1,2 @@
+import Controls from './Controls'
+export { Controls }
