+++ packages/@dcl/inspector/src/components/Input/Input.tsx
@@ -9,7 +9,7 @@ interface Props {
 const submittingKeys = new Set(['Enter'])
 const cancelingKeys = new Set(['Escape', 'Tab'])
 
-export const Input = ({ value, onCancel, onSubmit }: Props) => {
+const Input = ({ value, onCancel, onSubmit }: Props) => {
   const ref = useRef<HTMLInputElement>(null)
   const [stateValue, setStateValue] = useState(value)
 
@@ -46,3 +46,5 @@ export const Input = ({ value, onCancel, onSubmit }: Props) => {
 
   return <input ref={ref} type="text" value={stateValue} onChange={handleTextChange} />
 }
+
+export default React.memo(Input)
