+++ packages/@dcl/inspector/src/components/Input/index.ts
@@ -1,0 +1,2 @@
+import Input from './Input'
+export { Input }
